Move utility class to utils package
